agent/pyproject.toml
[project]
name = "pydantic-work"
version = "0.1.10"
description = "A local-first chat UI for Pydantic AI agents"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "David Sanchez", email = "64162682+dsfaccini@users.noreply.github.com" }
]
license = "MIT"
dependencies = [
    "bs4>=0.0.2",
    "fastapi>=0.117.1",
    "httpx>=0.27.0",
    "lancedb>=0.25.0",
    "langchain-text-splitters>=0.3.11",
    "logfire[fastapi]>=4.9.0",
    "markdown2>=2.5.4",
    "pip>=25.2",
    "pydantic-ai-slim[anthropic,openai,google,cli]>=1.0.10",
    "pyright>=1.1.405",
    "python-frontmatter>=1.1.0",
    "questionary>=2.0.0",
    "sentence-transformers>=5.1.1",
    "sse-starlette>=3.0.2",
    "starlette>=0.48.0",
    "uvicorn>=0.37.0",
    "watchfiles>=1.1.0",
]

[project.urls]
Homepage = "https://github.com/dsfaccini/ai-chat-ui/tree/pydantic-work"
Repository = "https://github.com/dsfaccini/ai-chat-ui/tree/pydantic-work"
Issues = "https://github.com/dsfaccini/ai-chat-ui/tree/pydantic-work/issues"
Twitter = "https://x.com/dasfacc"

[project.scripts]
pydantic-work = "chat.studio:main"

[tool.setuptools]
packages = ["chat"]

[tool.uv]
package = true

[tool.uv.sources]
pydantic-ai-slim = { git = "https://github.com/pydantic/pydantic-ai.git", subdirectory = "pydantic_ai_slim", rev = "main" }

[tool.ruff.format]
# don't format python in docstrings, pytest-examples takes care of it
docstring-code-format = false
quote-style = "single"
